Bill_Managing_Agent_Autogen/requirements.txt
pyautogen
Pillow 